+register & login

+++ controllers/cards.js
@@ -1,8 +1,10 @@
 const Card = require('../models/card');
+const NotFoundError = require('../errors/not-found-err');
+const ValidationError = require('../errors/validation-err');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
-    .orFail(new Error('NoValidCards'))
+    .orFail(new NotFoundError('Карточки не найдены.'))
     .then((cards) => res.send(cards.map(
       (card) => ({
         _id: card._id,
@@ -13,16 +15,10 @@ module.exports.getCards = (req, res) => {
         createdAt: card.createdAt,
       }),
     )))
-    .catch((err) => {
-      if (err.message === 'NoValidCards') {
-        res.status(404).send({ message: 'Карточки не найдены.' });
-      } else {
-        res.status(500).send({ message: 'Произошла ошибка' });
-      }
-    });
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
 
   Card.create({ name, link, owner: req.user._id })
@@ -36,43 +32,50 @@ module.exports.createCard = (req, res) => {
     }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(400).send({ message: 'Переданы некорректные данные при создании карточки.' });
-      } else {
-        res.status(500).send({ message: 'Произошла ошибка' });
+        throw new ValidationError('Переданы некорректные данные при создании карточки.');
       }
-    });
+    })
+    .catch(next);
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail(new Error('NoValidCard'))
-    .then((card) => res.send({
-      _id: card._id,
-      name: card.name,
-      link: card.link,
-      likes: card.likes,
-      owner: card.owner,
-      createdAt: card.createdAt,
-    }))
-    .catch((err) => {
-      if (err.message === 'NoValidCard') {
-        res.status(404).send({ message: 'Карточка с указанным _id не найдена.' });
-      } else if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Переданы некорректные данные _id.' });
-      } else {
-        res.status(500).send({ message: 'Произошла ошибка' });
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
+    .orFail(new NotFoundError('Карточка с указанным _id не найдена.'))
+    .then((cardCheck) => {
+      if (!(req.user._id === cardCheck.owner.toString())) {
+        throw new Error('NoRights');
       }
-    });
+      return Card.findByIdAndRemove(req.params.cardId)
+        .orFail(new NotFoundError('Карточка с указанным _id не найдена.'))
+        .then((card) => {
+          res.send({
+            _id: card._id,
+            name: card.name,
+            link: card.link,
+            likes: card.likes,
+            owner: card.owner,
+            createdAt: card.createdAt,
+          });
+        });
+    })
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        throw new ValidationError('Переданы некорректные данные _id.');
+      } else if (err.message === 'NoRights') {
+        res.status(403).send({ message: 'Недостаточно прав для удаления карточки.' });
+      }
+    })
+    .catch(next);
 };
 
 // поставить лайк карточке
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
     { new: true },
   )
-    .orFail(new Error('NoValidCard'))
+    .orFail(new NotFoundError('Карточка с указанным _id не найдена.'))
     .then((card) => res.send({
       _id: card._id,
       name: card.name,
@@ -82,24 +85,21 @@ module.exports.likeCard = (req, res) => {
       createdAt: card.createdAt,
     }))
     .catch((err) => {
-      if (err.message === 'NoValidCard') {
-        res.status(404).send({ message: 'Карточка с указанным _id не найдена.' });
-      } else if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Переданы некорректные данные _id карточки для постановки лайка.' });
-      } else {
-        res.status(500).send({ message: 'Произошла ошибка' });
+      if (err.name === 'CastError') {
+        throw new ValidationError('Переданы некорректные данные _id карточки для постановки лайка.');
       }
-    });
+    })
+    .catch(next);
 };
 
 // убрать лайк с карточки
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } }, // убрать _id из массива
     { new: true },
   )
-    .orFail(new Error('NoValidCard'))
+    .orFail(new NotFoundError('Карточка с указанным _id не найдена.'))
     .then((card) => res.send({
       _id: card._id,
       name: card.name,
@@ -109,12 +109,9 @@ module.exports.dislikeCard = (req, res) => {
       createdAt: card.createdAt,
     }))
     .catch((err) => {
-      if (err.message === 'NoValidCard') {
-        res.status(404).send({ message: 'Карточка с указанным _id не найдена.' });
-      } else if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Переданы некорректные данные _id карточки для снятии лайка.' });
-      } else {
-        res.status(500).send({ message: 'Произошла ошибка' });
+      if (err.name === 'CastError') {
+        throw new ValidationError('Переданы некорректные данные _id карточки для снятии лайка.');
       }
-    });
+    })
+    .catch(next);
 };
